requirements.txt
pyyaml>=6.0.1
requests>=2.32.3